lp-731 : delete GP - use optional - sq

+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/service/GeneralPartnerService.java
@@ -24,6 +24,7 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 import static uk.gov.companieshouse.limitedpartnershipsapi.utils.Constants.FILING_KIND_GENERAL_PARTNER;
@@ -194,7 +195,11 @@ public class GeneralPartnerService {
     }
 
     public void deleteGeneralPartner(Transaction transaction, String generalPartnerId, String requestId) throws ServiceException {
-        repository.findById(generalPartnerId).orElseThrow(() -> new ResourceNotFoundException(String.format("General partner with id %s not found", generalPartnerId)));
+        Optional<GeneralPartnerDao> generalPartnerDao = repository.findById(generalPartnerId);
+
+        if (generalPartnerDao.isEmpty()) {
+            throw new ResourceNotFoundException(String.format("General partner with id %s not found", generalPartnerId));
+        }
 
         repository.deleteById(generalPartnerId);
